requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
numpy
pillow
matplotlib
seaborn
scikit-learn
opencv-python
pandas
